+++ CHANGELOG.md
@@ -8,9 +8,10 @@
 * Fixed security groups, adding `-G` support
 * Added snapshots as a new column in image list
 * "knife openstack image list" fails with empty image name (KNIFE-83) (Simon Belluzzo)
+* excon / fog errors are a JSON blob, Rescue fog errors (KNIFE-87) (Bryan McLellan)
+* Better error handling for connection errors.
 
 TODO:
-* excon / fog errors are a JSON blob, Rescue fog errors (KNIFE-87) (Bryan McLellan)
 * Pass ssh_password to bootstrap (David Petzel) knife openstack server create doesn't pass along SSH Password (KNIFE-88)
 * Attach to floating IPs (Mohit Sethi)
 * Allow an option to ignore the SSL cert (KNIFE-225